requirements.txt
# Core dependencies - pinned to known-working versions
# Install with: pip install -r requirements.txt
#
# Note: PyAudio requires portaudio19-dev headers. See setup.sh.

# STT engine
RealtimeSTT==0.3.0
faster-whisper==1.0.3

# PyTorch (CUDA 12.1) - GPU acceleration
torch==2.3.1
torchaudio==2.3.1

# Wake word detection
openwakeword==0.6.0
onnxruntime==1.19.2
tflite-runtime==2.14.0

# Audio
PyAudio==0.2.14
webrtcvad==2.0.10

# Push-to-talk keyboard listener
pynput==1.8.1

# Voice commands config
PyYAML==6.0.3

# ── Agent mode (optional) ──────────────────────────────────
# These are only needed if you run with --agent flag.
# Requires Python >=3.10 (for qwen-tts / accelerate).

# LLM client (works with Ollama, Groq, OpenAI - all OpenAI-compatible)
openai>=1.0.0

# Local TTS — Qwen3-TTS (GPU-accelerated, fully offline)
qwen-tts>=0.1.0
soundfile

# Cloud TTS (optional - only if using ElevenLabs)
# elevenlabs
